feat: update c sdk implement.
