+++ sdk/c/Cargo.toml
@@ -22,7 +22,9 @@ doc = false
 
 [dependencies]
 datenlord = { path = "../../" }
+clap = { version = "4", features = ["derive"] }
 bytes = "1.6.0"
+parking_lot = "0.12.0"
 tokio = { version = "1.32.0", features = ["full"] }
 tokio-util = "0.7.10"
 async-trait = "0.1.48"
