+++ sdk/c/examples/test_datenlord_sdk.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
-#include "datenlord.h"
+#include "datenlordsdk.h"
 
 
 // Utils functions to free current error
